requirements.txt
streamlit==1.37.0
opencv-python-headless==4.10.0.82
numpy==1.26.4
Pillow==10.3.0
